backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
black==24.10.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
dj-rest-auth==2.2.6
Django==4.2.20
django-allauth==65.1.0
django-cors-headers==4.5.0
django-health-check==3.18.3
django-oauth-toolkit==3.0.1
django-stubs==5.1.0
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.17.1
elasticsearch==8.17.1
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
importlib_resources==6.5.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
kiwisolver==1.4.7
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.1.147
matplotlib==3.9.0
multidict==6.4.3
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.82.0
orjson==3.10.16
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
pgvector==0.3.6
pillow==11.2.1
platformdirs==4.3.7
pre_commit==4.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
types-PyYAML==6.0.12.20250402
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
yarl==1.20.0
zipp==3.21.0
